Move `entity_db_button_ui` and make it usable for blueprints too
